WIP on fmi-check
